requirements.txt
requests
python-dotenv

pillow # for image processing